// app/src/generated/Eval.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Side } from "./Side";

/**
 * A position evaluation from White's point of view.
 */
export type Eval = { "kind": "cp", "value": number } | { "kind": "mate", "value": number } | { "kind": "checkmate", "value": Side };